requirements.txt
Django==3.2.7
django-cors-middleware==1.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0